Update currentTime value in Video.js cache
fixes #43

The source of the problem is: SeekBar.getCurrentTime_() retrieves the current time from the cache when the player is in scrubbling mode (i.e. while the mouse button is down). The cached currentTime value is set at the end of Player.currentTime(). The videojs-offset plugin overrides Player.currentTime() and modifies the returned value, but the original (wrong) currentTime value gets cached. This patch fixes that by updating the cached value.